refactor: rename echo resolvers to store resolvers, split out createRequestBodyEcho
