+++ src/msw/api/businesses/[business-id]/custom-accounts/post.ts
@@ -5,8 +5,8 @@ import { type CustomAccount, CustomAccountSchema } from '@schemas/customAccounts
 import { customAccountFromCreateRequest } from '@msw/api/businesses/[business-id]/custom-accounts/customAccountFromCreateRequest'
 import { customAccountStore } from '@msw/api/businesses/[business-id]/custom-accounts/store'
 import { apiData } from '@msw/utils/apiResponse'
-import { createEchoCreateResolver } from '@msw/utils/createEchoResolvers'
 import { createMockEndpoint } from '@msw/utils/createMockEndpoint'
+import { createStoreCreateResolver } from '@msw/utils/createStoreResolvers'
 import { makeCustomAccount } from '@fixtures/customAccounts/mocks'
 
 const encodeCustomAccount = Schema.encodeSync(CustomAccountSchema)
@@ -17,7 +17,7 @@ export const toCreateCustomAccountResponse = (customAccount: CustomAccount) =>
 export const post = createMockEndpoint<CustomAccount, ReturnType<typeof toCreateCustomAccountResponse>>({
   method: 'post',
   path: '*/v1/businesses/:businessId/custom-accounts',
-  resolve: createEchoCreateResolver({
+  resolve: createStoreCreateResolver({
     store: customAccountStore,
     makeBase: id => makeCustomAccount({ id }),
     fromRequest: customAccountFromCreateRequest,
