+++ src/msw/api/businesses/[business-id]/customers/patch.ts
@@ -5,8 +5,8 @@ import { type Customer, CustomerSchema } from '@schemas/customer'
 import { customerFromUpsertRequest } from '@msw/api/businesses/[business-id]/customers/customerFromUpsertRequest'
 import { customerStore } from '@msw/api/businesses/[business-id]/customers/store'
 import { apiData } from '@msw/utils/apiResponse'
-import { createEchoUpdateResolver } from '@msw/utils/createEchoResolvers'
 import { createMockEndpoint } from '@msw/utils/createMockEndpoint'
+import { createStoreUpdateResolver } from '@msw/utils/createStoreResolvers'
 import { makeCustomer } from '@fixtures/customers/mocks'
 
 const encodeCustomer = Schema.encodeSync(CustomerSchema)
@@ -17,7 +17,7 @@ export const toUpdateCustomerResponse = (customer: Customer) =>
 export const patch = createMockEndpoint<Customer, ReturnType<typeof toUpdateCustomerResponse>>({
   method: 'patch',
   path: '*/v1/businesses/:businessId/customers/:customerId',
-  resolve: createEchoUpdateResolver({
+  resolve: createStoreUpdateResolver({
     idParam: 'customerId',
     store: customerStore,
     makeBase: id => makeCustomer({ id }),
