+++ src/msw/api/businesses/[business-id]/customers/post.ts
@@ -5,8 +5,8 @@ import { type Customer, CustomerSchema } from '@schemas/customer'
 import { customerFromUpsertRequest } from '@msw/api/businesses/[business-id]/customers/customerFromUpsertRequest'
 import { customerStore } from '@msw/api/businesses/[business-id]/customers/store'
 import { apiData } from '@msw/utils/apiResponse'
-import { createEchoCreateResolver } from '@msw/utils/createEchoResolvers'
 import { createMockEndpoint } from '@msw/utils/createMockEndpoint'
+import { createStoreCreateResolver } from '@msw/utils/createStoreResolvers'
 import { makeCustomer } from '@fixtures/customers/mocks'
 
 const encodeCustomer = Schema.encodeSync(CustomerSchema)
@@ -17,7 +17,7 @@ export const toCreateCustomerResponse = (customer: Customer) =>
 export const post = createMockEndpoint<Customer, ReturnType<typeof toCreateCustomerResponse>>({
   method: 'post',
   path: '*/v1/businesses/:businessId/customers',
-  resolve: createEchoCreateResolver({
+  resolve: createStoreCreateResolver({
     store: customerStore,
     makeBase: id => makeCustomer({ id }),
     fromRequest: customerFromUpsertRequest,
