+++ src/msw/api/businesses/[business-id]/mileage/trips/patch.ts
@@ -5,8 +5,8 @@ import { type Trip, TripSchema } from '@schemas/trip'
 import { tripStore } from '@msw/api/businesses/[business-id]/mileage/trips/store'
 import { tripFromUpsertRequest } from '@msw/api/businesses/[business-id]/mileage/trips/tripFromUpsertRequest'
 import { apiData } from '@msw/utils/apiResponse'
-import { createEchoUpdateResolver } from '@msw/utils/createEchoResolvers'
 import { createMockEndpoint } from '@msw/utils/createMockEndpoint'
+import { createStoreUpdateResolver } from '@msw/utils/createStoreResolvers'
 import { makeTrip } from '@fixtures/trips/mocks'
 
 const encodeTrip = Schema.encodeSync(TripSchema)
@@ -17,7 +17,7 @@ export const toUpdateTripResponse = (trip: Trip) =>
 export const patch = createMockEndpoint<Trip, ReturnType<typeof toUpdateTripResponse>>({
   method: 'patch',
   path: '*/v1/businesses/:businessId/mileage/trips/:tripId',
-  resolve: createEchoUpdateResolver({
+  resolve: createStoreUpdateResolver({
     idParam: 'tripId',
     store: tripStore,
     makeBase: id => makeTrip({ id }),
