+++ src/msw/api/businesses/[business-id]/mileage/trips/post.ts
@@ -5,8 +5,8 @@ import { type Trip, TripSchema } from '@schemas/trip'
 import { tripStore } from '@msw/api/businesses/[business-id]/mileage/trips/store'
 import { tripFromUpsertRequest } from '@msw/api/businesses/[business-id]/mileage/trips/tripFromUpsertRequest'
 import { apiData } from '@msw/utils/apiResponse'
-import { createEchoCreateResolver } from '@msw/utils/createEchoResolvers'
 import { createMockEndpoint } from '@msw/utils/createMockEndpoint'
+import { createStoreCreateResolver } from '@msw/utils/createStoreResolvers'
 import { makeTrip } from '@fixtures/trips/mocks'
 
 const encodeTrip = Schema.encodeSync(TripSchema)
@@ -17,7 +17,7 @@ export const toCreateTripResponse = (trip: Trip) =>
 export const post = createMockEndpoint<Trip, ReturnType<typeof toCreateTripResponse>>({
   method: 'post',
   path: '*/v1/businesses/:businessId/mileage/trips',
-  resolve: createEchoCreateResolver({
+  resolve: createStoreCreateResolver({
     store: tripStore,
     makeBase: id => makeTrip({ id }),
     fromRequest: tripFromUpsertRequest,
