+++ src/msw/api/businesses/[business-id]/mileage/vehicles/[vehicle-id]/archive/post.ts
@@ -4,8 +4,8 @@ import { type Vehicle, VehicleSchema } from '@schemas/vehicle'
 
 import { vehicleStore } from '@msw/api/businesses/[business-id]/mileage/vehicles/store'
 import { apiData } from '@msw/utils/apiResponse'
-import { createEchoTransformationResolver } from '@msw/utils/createEchoResolvers'
 import { createMockEndpoint } from '@msw/utils/createMockEndpoint'
+import { createStoreTransformResolver } from '@msw/utils/createStoreResolvers'
 import { makeVehicle } from '@fixtures/vehicles/mocks'
 
 const encodeVehicle = Schema.encodeSync(VehicleSchema)
@@ -16,7 +16,7 @@ export const toArchiveVehicleResponse = (vehicle: Vehicle) =>
 export const post = createMockEndpoint<Vehicle, ReturnType<typeof toArchiveVehicleResponse>>({
   method: 'post',
   path: '*/v1/businesses/:businessId/mileage/vehicles/:vehicleId/archive',
-  resolve: createEchoTransformationResolver({
+  resolve: createStoreTransformResolver({
     idParam: 'vehicleId',
     store: vehicleStore,
     makeBase: id => makeVehicle({ id }),
