+++ src/msw/api/businesses/[business-id]/mileage/vehicles/patch.ts
@@ -5,8 +5,8 @@ import { type Vehicle, VehicleSchema } from '@schemas/vehicle'
 import { vehicleStore } from '@msw/api/businesses/[business-id]/mileage/vehicles/store'
 import { vehicleFromUpsertRequest } from '@msw/api/businesses/[business-id]/mileage/vehicles/vehicleFromUpsertRequest'
 import { apiData } from '@msw/utils/apiResponse'
-import { createEchoUpdateResolver } from '@msw/utils/createEchoResolvers'
 import { createMockEndpoint } from '@msw/utils/createMockEndpoint'
+import { createStoreUpdateResolver } from '@msw/utils/createStoreResolvers'
 import { makeVehicle } from '@fixtures/vehicles/mocks'
 
 const encodeVehicle = Schema.encodeSync(VehicleSchema)
@@ -17,7 +17,7 @@ export const toUpdateVehicleResponse = (vehicle: Vehicle) =>
 export const patch = createMockEndpoint<Vehicle, ReturnType<typeof toUpdateVehicleResponse>>({
   method: 'patch',
   path: '*/v1/businesses/:businessId/mileage/vehicles/:vehicleId',
-  resolve: createEchoUpdateResolver({
+  resolve: createStoreUpdateResolver({
     idParam: 'vehicleId',
     store: vehicleStore,
     makeBase: id => makeVehicle({ id }),
