+++ src/msw/api/businesses/[business-id]/mileage/vehicles/post.ts
@@ -5,8 +5,8 @@ import { type Vehicle, VehicleSchema } from '@schemas/vehicle'
 import { vehicleStore } from '@msw/api/businesses/[business-id]/mileage/vehicles/store'
 import { vehicleFromUpsertRequest } from '@msw/api/businesses/[business-id]/mileage/vehicles/vehicleFromUpsertRequest'
 import { apiData } from '@msw/utils/apiResponse'
-import { createEchoCreateResolver } from '@msw/utils/createEchoResolvers'
 import { createMockEndpoint } from '@msw/utils/createMockEndpoint'
+import { createStoreCreateResolver } from '@msw/utils/createStoreResolvers'
 import { makeVehicle } from '@fixtures/vehicles/mocks'
 
 const encodeVehicle = Schema.encodeSync(VehicleSchema)
@@ -17,7 +17,7 @@ export const toCreateVehicleResponse = (vehicle: Vehicle) =>
 export const post = createMockEndpoint<Vehicle, ReturnType<typeof toCreateVehicleResponse>>({
   method: 'post',
   path: '*/v1/businesses/:businessId/mileage/vehicles',
-  resolve: createEchoCreateResolver({
+  resolve: createStoreCreateResolver({
     store: vehicleStore,
     makeBase: id => makeVehicle({ id }),
     fromRequest: vehicleFromUpsertRequest,
